+++ src/main/java/roomescape/reservation/domain/entity/MemberReservation.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.domain;
+package roomescape.reservation.domain.entity;
 
 import jakarta.persistence.*;
 import org.hibernate.annotations.ColumnDefault;
