+++ src/test/java/roomescape/Fixtures.java
@@ -2,8 +2,8 @@ package roomescape;
 
 import roomescape.auth.dto.LoginMember;
 import roomescape.member.domain.Member;
-import roomescape.reservation.domain.MemberReservation;
-import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.domain.entity.MemberReservation;
+import roomescape.reservation.domain.entity.Reservation;
 import roomescape.time.domain.ReservationTime;
 import roomescape.theme.domain.Theme;
